--- packages/manager/src/components/PaginationControls/index.tsx
@@ -1,2 +1,0 @@
-import PaginationControls from './PaginationControls';
-export default PaginationControls;
